+++ src/trayicon.cpp
@@ -81,7 +81,7 @@ void TrayIcon::processTrayActivated(QSystemTrayIcon::ActivationReason reason)
         return;
 
     auto *mainWindow = qobject_cast<MainWindow *>(parent());
-    if (!mainWindow->isVisible())
+    if (!mainWindow->isVisible() || mainWindow->isMinimized())
         mainWindow->open();
     else
         mainWindow->hide();